executable-demo/api/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
psycopg2-binary
